+++ src/main/java/de/r3s6/maven/constcreator/PropEntry.java
@@ -15,25 +15,27 @@ package de.r3s6.maven.constcreator;
  * limitations under the License.
  */
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Objects;
+
+import de.r3s6.maven.constcreator.NameHandler.JavaNames;
 
 /**
  * PropEntry describes a property file entry with generated Java names.
  * <p>
  * An instance of this class holds the key and value from the properties file
  * and additional the generated constant, variable and getter name.
+ * <p>
+ * Also escaped key and value for javadoc are available.
  *
  * @author Ralf Schandl
  */
 public class PropEntry {
 
-    private static final String DIV = "_";
-    private static final char DIV_CHR = DIV.charAt(0);
-
     private final String key;
     private final String value;
+
+    private final String javadocKey;
+    private final String javadocValue;
 
     private final String constantName;
     private final String variableName;
@@ -57,11 +59,23 @@ public class PropEntry {
         }
         this.value = value;
 
-        final List<String> parts = splitParts(key);
+        this.javadocKey = escapeJavadoc(this.key);
+        this.javadocValue = escapeJavadoc(this.value);
 
-        constantName = buildConstantName(parts);
-        variableName = buildVariableName(parts);
-        getterName = buildGetterName(parts);
+        final JavaNames names = NameHandler.createJavaNames(key);
+        constantName = names.getConstantName();
+        variableName = names.getVariableName();
+        getterName = names.getGetterName();
+    }
+
+    private String escapeJavadoc(final String text) {
+        // @formatter:off
+        return text.replace("&", "&amp;")
+                .replace("<", "&lt;")
+                .replace(">", "&gt;")
+                .replace("{@", "{&#64;")
+                .replace("*/", "&#42;/");
+        // @formatter:on
     }
 
     public String getConstantName() {
@@ -84,76 +98,11 @@ public class PropEntry {
         return value;
     }
 
-    private List<String> splitParts(final String propKey) {
-        final List<String> parts = new ArrayList<>();
-
-        final StringBuilder sb = new StringBuilder();
-
-        boolean lastIsLower = false;
-
-        for (char chr : propKey.toCharArray()) {
-
-            final boolean isUpper = Character.isUpperCase(chr);
-
-            if (chr != DIV_CHR && Character.isJavaIdentifierPart(chr)) {
-                if (lastIsLower && isUpper) {
-                    parts.add(sb.toString());
-                    sb.setLength(0);
-                }
-                sb.append(chr);
-                lastIsLower = !isUpper;
-            } else if (sb.length() > 0) {
-                parts.add(sb.toString());
-                sb.setLength(0);
-                lastIsLower = false;
-            }
-        }
-
-        if (sb.length() > 0) {
-            parts.add(sb.toString());
-        }
-
-        return parts;
-
+    public String getJavadocKey() {
+        return javadocKey;
     }
 
-    private String buildConstantName(final List<String> parts) {
-        final String name = String.join(DIV, parts).toUpperCase();
-        if (Character.isJavaIdentifierStart(name.charAt(0))) {
-            return name;
-        } else {
-            return DIV + name;
-        }
-
+    public String getJavadocValue() {
+        return javadocValue;
     }
-
-    private String buildVariableName(final List<String> parts) {
-
-        final StringBuilder sb = new StringBuilder();
-
-        for (String part : parts) {
-            sb.append(Character.toUpperCase(part.charAt(0)));
-            sb.append(part.substring(1).toLowerCase());
-        }
-
-        if (Character.isJavaIdentifierStart(sb.charAt(0))) {
-            sb.setCharAt(0, Character.toLowerCase(sb.charAt(0)));
-        } else {
-            sb.insert(0, DIV);
-        }
-
-        return sb.toString();
-    }
-
-    private String buildGetterName(final List<String> parts) {
-        final StringBuilder sb = new StringBuilder("get");
-
-        for (String part : parts) {
-            sb.append(Character.toUpperCase(part.charAt(0)));
-            sb.append(part.substring(1).toLowerCase());
-        }
-
-        return sb.toString();
-    }
-
 }