+++ src/test/resources-its/de/r3s6/maven/constcreator/inttest/ProjectBuildIT/testProject/src/test/java/de/tester/constants/messages/MessagesTest.java
@@ -34,7 +34,7 @@ class MessagesTest {
 
     @Test
     void testPropertiesFilenameName() {
-        assertEquals("messages/messages", Messages.BUNDLE_NAME);
+        assertEquals("messages.messages", Messages.BUNDLE_NAME);
     }
 
     @Test